edit now works with better logic

+++ app/src/main/java/com/example/piller/models/Dose.kt
@@ -8,5 +8,5 @@ import kotlinx.android.parcel.Parcelize
 class Dose(
     @SerializedName("dose_id") var doseId: String = "",
     @SerializedName("measurement_type") var measurementType: String = "",
-    @SerializedName("total_dose") var totalDose: Int = 0
+    @SerializedName("total_dose") var totalDose: Float = 0F
 ) : Parcelable
